requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
astor==0.8.1
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
conformer==0.3.2
contourpy==1.3.0
cosyvoice==0.0.5
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
easyocr==1.7.2
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
gdown==5.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.36.1
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.5.1
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.4
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
ninja==1.11.1.3
numba==0.60.0
numpy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_dbvs82799s/croot/numpy_and_numpy_base_1677549685398/work
omegaconf==2.3.0
onnxruntime==1.19.2
openai-whisper==20240930
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pillow==11.0.0
propcache==0.2.1
protobuf==5.29.2
pyclipper==1.3.0.post6
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.0
PySocks==1.7.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
RapidFuzz==3.11.0
regex==2024.11.6
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.13.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.8.30
tiktoken==0.8.0
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
typeguard==4.4.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
wget==3.2
yarl==1.18.3
zipp==3.21.0
